refactor(#112): 카테고리 로직 분리
