+++ src/main/java/deepdivers/community/domain/category/CategoryExceptionCode.java
@@ -1,4 +1,4 @@
-package deepdivers.community.domain.post.exception;
+package deepdivers.community.domain.category;
 
 import deepdivers.community.domain.common.dto.code.ExceptionCode;
 import lombok.AllArgsConstructor;
