+++ src/main/java/deepdivers/community/domain/category/CategoryRepository.java
@@ -1,8 +1,8 @@
-package deepdivers.community.domain.post.repository.jpa;
+package deepdivers.community.domain.category;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import deepdivers.community.domain.post.entity.PostCategory;
+import deepdivers.community.domain.category.entity.PostCategory;
 
 public interface CategoryRepository extends JpaRepository<PostCategory, Long> {
 }
