+++ src/main/java/deepdivers/community/domain/category/entity/PostCategory.java
@@ -1,9 +1,7 @@
-package deepdivers.community.domain.post.entity;
+package deepdivers.community.domain.category.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
