+++ src/test/java/deepdivers/community/domain/category/entity/PostCategoryTest.java
@@ -1,12 +1,9 @@
-package deepdivers.community.domain.post.entity;
+package deepdivers.community.domain.category.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.Arrays;
-import java.util.List;
 
 class PostCategoryTest {
